上传照片修复

+++ wechat/controllers/MemberController.php
@@ -93,20 +93,25 @@ class MemberController extends BaseController
     public function actionSavePhoto()
     {
         $user_id = Cookie::getInstance()->getCookie('bhy_id')->value;
+        $list = [];
         foreach($this->get as $k => $v) {
             $arr = json_decode($v);
+            $thuArr = explode('/', $arr->thumb_path);
+            $picArr = explode('/', $arr->pic_path);
             $data[$k]['type'] = $arr->type;
             $data[$k]['pic_path'] = str_replace('thumb', 'picture', $arr->thumb_path);
             $data[$k]['thumb_path'] = $arr->thumb_path;
-            // 删除原有图片
-            $pic_path = __DIR__ . "/../.." . $arr->pic_path;
-            if (is_file($pic_path) && unlink($pic_path)) {
-                $thumb_path = str_replace('picture', 'thumb', $pic_path);
-                unlink($thumb_path);
+            if($thuArr[5] != $picArr[5]) {
+                // 删除原有图片
+                $pic_path = __DIR__ . "/../.." . $arr->pic_path;
+                if (is_file($pic_path) && unlink($pic_path)) {
+                    $thumb_path = str_replace('picture', 'thumb', $pic_path);
+                    unlink($thumb_path);
+                }
             }
+            $list = UserPhoto::getInstance()->savePhoto($data, $user_id);
         }
-        $list = UserPhoto::getInstance()->savePhoto($data, $user_id);
-        $this->renderAjax(['status' => 1, 'data' => $list]);
+            $this->renderAjax(['status' => 1, 'data' => $list]);
     }
 
     /**
